+++ src/pages/iv-categories/index.js
@@ -2,28 +2,22 @@ import { useContext, useState, useEffect } from 'react'
 import toast from 'react-hot-toast'
 import Table from 'src/components/Shared/Table'
 import { RequestsContext } from 'src/providers/RequestsContext'
-import { Button, Grid, Tooltip } from '@mui/material'
 import { useWindow } from 'src/windows'
 import { useResourceQuery } from 'src/hooks/resource'
 import { ResourceIds } from 'src/resources/ResourceIds'
 import { VertLayout } from 'src/components/Shared/Layouts/VertLayout'
 import { Fixed } from 'src/components/Shared/Layouts/Fixed'
 import { Grow } from 'src/components/Shared/Layouts/Grow'
-import PreviewReport from 'src/components/Shared/PreviewReport'
 import Tree from 'src/components/Shared/Tree'
 import { ControlContext } from 'src/providers/ControlContext'
 import GridToolbar from 'src/components/Shared/GridToolbar'
 import { InventoryRepository } from 'src/repositories/InventoryRepository'
-import { SystemRepository } from 'src/repositories/SystemRepository'
 import CategoryWindow from './window/CategoryWindow'
-import CustomComboBox from 'src/components/Inputs/CustomComboBox'
 
 const Category = () => {
   const { getRequest, postRequest } = useContext(RequestsContext)
   const { platformLabels } = useContext(ControlContext)
   const { stack } = useWindow()
-  const [selectedReport, setSelectedReport] = useState(null)
-  const [reportStore, setReportStore] = useState([])
   const [dataTree, setDataTree] = useState([])
 
   const {
@@ -141,41 +135,6 @@ const Category = () => {
     openForm(obj?.recordId)
   }
 
-  useEffect(() => {
-    getReportLayout()
-  }, [])
-
-  useEffect(() => {
-    if (reportStore.length > 0) {
-      setSelectedReport(reportStore[0])
-    } else {
-      setSelectedReport(null)
-    }
-  }, [reportStore])
-
-  const getReportLayout = () => {
-    setReportStore([])
-
-    getRequest({
-      extension: SystemRepository.ReportLayout,
-      parameters: `_resourceId=${ResourceIds.Category}`
-    }).then(res => {
-      if (res?.list) {
-        const formattedReports = res.list.map(item => ({
-          api_url: item.api,
-          reportClass: item.instanceName,
-          parameters: item.parameters,
-          layoutName: item.layoutName,
-          assembly: 'ArgusRPT.dll'
-        }))
-        setReportStore(formattedReports)
-        if (formattedReports.length > 0) {
-          setSelectedReport(formattedReports[0])
-        }
-      }
-    })
-  }
-
   function onTreeClick() {
     stack({
       Component: Tree,
@@ -212,52 +171,7 @@ const Category = () => {
           onTree={onTreeClick}
           onAdd={add}
           maxAccess={access}
-          rightSection={
-            <Grid item sx={{ display: 'flex', mr: 2 }}>
-              <CustomComboBox
-                label={platformLabels.SelectReport}
-                valueField='caption'
-                displayField='layoutName'
-                store={reportStore}
-                value={selectedReport}
-                onChange={(e, newValue) => setSelectedReport(newValue)}
-                sx={{ width: 250 }}
-                disableClearable
-              />
-              <Button
-                variant='contained'
-                disabled={!selectedReport}
-                onClick={() =>
-                  stack({
-                    Component: PreviewReport,
-                    props: {
-                      selectedReport: selectedReport,
-                      outerGrid: true
-                    },
-                    width: 1000,
-                    height: 500,
-                    title: platformLabels.PreviewReport
-                  })
-                }
-                sx={{
-                  ml: 2,
-                  backgroundColor: '#231F20',
-                  '&:hover': {
-                    backgroundColor: '#231F20',
-                    opacity: 0.8
-                  },
-                  width: 'auto',
-                  height: '35px',
-                  objectFit: 'contain'
-                }}
-                size='small'
-              >
-                <Tooltip title={platformLabels.Preview}>
-                  <img src='/images/buttonsIcons/preview.png' alt={platformLabels.Preview} />
-                </Tooltip>
-              </Button>
-            </Grid>
-          }
+          previewReport={ResourceIds.Category}
         />
       </Fixed>
       <Grow>
