+++ src/pages/iv-item-groups/index.js
@@ -1,8 +1,5 @@
-import { useContext, useState, useEffect } from 'react'
+import { useContext } from 'react'
 import toast from 'react-hot-toast'
-import { Button, Grid, Tooltip } from '@mui/material'
-import CustomComboBox from 'src/components/Inputs/CustomComboBox'
-import PreviewReport from 'src/components/Shared/PreviewReport'
 import Table from 'src/components/Shared/Table'
 import GridToolbar from 'src/components/Shared/GridToolbar'
 import { RequestsContext } from 'src/providers/RequestsContext'
@@ -15,14 +12,11 @@ import { Grow } from 'src/components/Shared/Layouts/Grow'
 import { ControlContext } from 'src/providers/ControlContext'
 import IvItemGroupsForm from './form/IvItemGroupsForm'
 import { InventoryRepository } from 'src/repositories/InventoryRepository'
-import { SystemRepository } from 'src/repositories/SystemRepository'
 
 const IvItemGroups = () => {
   const { getRequest, postRequest } = useContext(RequestsContext)
   const { platformLabels } = useContext(ControlContext)
   const { stack } = useWindow()
-  const [selectedReport, setSelectedReport] = useState(null)
-  const [reportStore, setReportStore] = useState([])
 
   const {
     query: { data },
@@ -110,45 +104,6 @@ const IvItemGroups = () => {
     openForm(obj?.recordId)
   }
 
-  useEffect(() => {
-    getReportLayout()
-  }, [ResourceIds.InventoryGroup])
-
-  useEffect(() => {
-    if (reportStore.length > 0) {
-      setSelectedReport(reportStore[0])
-    } else {
-      setSelectedReport(null)
-    }
-  }, [reportStore])
-
-  const getReportLayout = () => {
-    setReportStore([])
-    if (ResourceIds.InventoryGroup) {
-      var parameters = `_resourceId=${ResourceIds.InventoryGroup}`
-      getRequest({
-        extension: SystemRepository.ReportLayout,
-        parameters: parameters
-      })
-        .then(res => {
-          if (res?.list) {
-            const formattedReports = res.list.map(item => ({
-              api_url: item.api,
-              reportClass: item.instanceName,
-              parameters: item.parameters,
-              layoutName: item.layoutName,
-              assembly: 'ArgusRPT.dll'
-            }))
-            setReportStore(formattedReports)
-            if (formattedReports.length > 0) {
-              setSelectedReport(formattedReports[0])
-            }
-          }
-        })
-        .catch(error => {})
-    }
-  }
-
   return (
     <VertLayout>
       <Fixed>
@@ -162,52 +117,7 @@ const IvItemGroups = () => {
           inputSearch={true}
           onAdd={add}
           maxAccess={access}
-          rightSection={
-            <Grid item sx={{ display: 'flex', mr: 2 }}>
-              <CustomComboBox
-                label={platformLabels.SelectReport}
-                valueField='caption'
-                displayField='layoutName'
-                store={reportStore}
-                value={selectedReport}
-                onChange={(e, newValue) => setSelectedReport(newValue)}
-                sx={{ width: 250 }}
-                disableClearable
-              />
-              <Button
-                variant='contained'
-                disabled={!selectedReport}
-                onClick={() =>
-                  stack({
-                    Component: PreviewReport,
-                    props: {
-                      selectedReport: selectedReport,
-                      outerGrid: true
-                    },
-                    width: 1000,
-                    height: 500,
-                    title: platformLabels.PreviewReport
-                  })
-                }
-                sx={{
-                  ml: 2,
-                  backgroundColor: '#231F20',
-                  '&:hover': {
-                    backgroundColor: '#231F20',
-                    opacity: 0.8
-                  },
-                  width: 'auto',
-                  height: '35px',
-                  objectFit: 'contain'
-                }}
-                size='small'
-              >
-                <Tooltip title={platformLabels.Preview}>
-                  <img src='/images/buttonsIcons/preview.png' alt={platformLabels.Preview} />
-                </Tooltip>
-              </Button>
-            </Grid>
-          }
+          previewReport={ResourceIds.InventoryGroup}
         />
       </Fixed>
       <Grow>
