+++ src/pages/mf-labors/index.js
@@ -105,7 +105,7 @@ const Labor = () => {
   return (
     <VertLayout>
       <Fixed>
-        <GridToolbar onAdd={add} maxAccess={access} />
+        <GridToolbar onAdd={add} maxAccess={access} previewReport={ResourceIds.Labor} />
       </Fixed>
       <Grow>
         <Table
