+++ src/pages/mf-operations/index.js
@@ -99,7 +99,7 @@ const Operationss = () => {
   return (
     <VertLayout>
       <Fixed>
-        <GridToolbar onAdd={add} maxAccess={access} />
+        <GridToolbar onAdd={add} maxAccess={access} previewReport={ResourceIds.Operations} />
       </Fixed>
       <Grow>
         <Table
