+++ src/pages/sites/index.js
@@ -1,5 +1,4 @@
-import { useState, useContext, useEffect } from 'react'
-import { Button, Grid, Tooltip } from '@mui/material'
+import { useContext } from 'react'
 import toast from 'react-hot-toast'
 import Table from 'src/components/Shared/Table'
 import GridToolbar from 'src/components/Shared/GridToolbar'
@@ -12,17 +11,10 @@ import SitesForm from './forms/SitesForm'
 import { VertLayout } from 'src/components/Shared/Layouts/VertLayout'
 import { Fixed } from 'src/components/Shared/Layouts/Fixed'
 import { Grow } from 'src/components/Shared/Layouts/Grow'
-import { ControlContext } from 'src/providers/ControlContext'
-import CustomComboBox from 'src/components/Inputs/CustomComboBox'
-import { SystemRepository } from 'src/repositories/SystemRepository'
-import PreviewReport from 'src/components/Shared/PreviewReport'
 
 const Sites = () => {
   const { getRequest, postRequest } = useContext(RequestsContext)
   const { stack } = useWindow()
-  const { platformLabels } = useContext(ControlContext)
-  const [selectedReport, setSelectedReport] = useState(null)
-  const [reportStore, setReportStore] = useState([])
 
   async function fetchGridData(options = {}) {
     const { _startAt = 0, _pageSize = 50 } = options
@@ -122,45 +114,6 @@ const Sites = () => {
     })
   }
 
-  useEffect(() => {
-    getReportLayout()
-  }, [ResourceIds.Sites])
-
-  useEffect(() => {
-    if (reportStore.length > 0) {
-      setSelectedReport(reportStore[0])
-    } else {
-      setSelectedReport(null)
-    }
-  }, [reportStore])
-
-  const getReportLayout = () => {
-    setReportStore([])
-    if (ResourceIds.Sites) {
-      var parameters = `_resourceId=${ResourceIds.Sites}`
-      getRequest({
-        extension: SystemRepository.ReportLayout,
-        parameters: parameters
-      })
-        .then(res => {
-          if (res?.list) {
-            const formattedReports = res.list.map(item => ({
-              api_url: item.api,
-              reportClass: item.instanceName,
-              parameters: item.parameters,
-              layoutName: item.layoutName,
-              assembly: 'ArgusRPT.dll'
-            }))
-            setReportStore(formattedReports)
-            if (formattedReports.length > 0) {
-              setSelectedReport(formattedReports[0])
-            }
-          }
-        })
-        .catch(error => {})
-    }
-  }
-
   return (
     <VertLayout>
       <Fixed>
@@ -172,52 +125,6 @@ const Sites = () => {
           labels={_labels}
           inputSearch={true}
           previewReport={ResourceIds.Sites}
-          rightSection={
-            <Grid item sx={{ display: 'flex', mr: 2 }}>
-              <CustomComboBox
-                label={platformLabels.SelectReport}
-                valueField='caption'
-                displayField='layoutName'
-                store={reportStore}
-                value={selectedReport}
-                onChange={(e, newValue) => setSelectedReport(newValue)}
-                sx={{ width: 250 }}
-                disableClearable
-              />
-              <Button
-                variant='contained'
-                disabled={!selectedReport}
-                onClick={() =>
-                  stack({
-                    Component: PreviewReport,
-                    props: {
-                      selectedReport: selectedReport,
-                      outerGrid: true
-                    },
-                    width: 1000,
-                    height: 500,
-                    title: platformLabels.PreviewReport
-                  })
-                }
-                sx={{
-                  ml: 2,
-                  backgroundColor: '#231F20',
-                  '&:hover': {
-                    backgroundColor: '#231F20',
-                    opacity: 0.8
-                  },
-                  width: 'auto',
-                  height: '35px',
-                  objectFit: 'contain'
-                }}
-                size='small'
-              >
-                <Tooltip title={platformLabels.Preview}>
-                  <img src='/images/buttonsIcons/preview.png' alt={platformLabels.Preview} />
-                </Tooltip>
-              </Button>
-            </Grid>
-          }
         />
       </Fixed>
       <Grow>
